refactor(tools): move credential/http_security/allowlist to ungated modules
These modules don't depend on wasmtime, so they live in src/tools/
directly instead of under src/tools/wasm/. Wasm-specific helpers like
reject_private_ip (uses blocking DNS for WASM host functions) remain
in src/tools/wasm/http_security.rs.

src/tools/wasm/mod.rs keeps public re-exports for backward compatibility
with existing caller import paths.
