+++ src/channels/wasm/wrapper.rs
@@ -54,7 +54,7 @@ use crate::channels::{Channel, IncomingMessage, MessageStream, OutgoingResponse,
 use crate::error::ChannelError;
 use crate::pairing::PairingStore;
 use crate::secrets::SecretsStore;
-use crate::tools::wasm::credential_injector::{
+use crate::tools::credentials::{
     InjectedCredentials, host_matches_pattern, inject_credential,
 };
 use crate::tools::wasm::{
