+++ src/tools/wasm/wrapper.rs
@@ -23,7 +23,7 @@ use crate::llm::recording::{HttpExchangeRequest, HttpExchangeResponse, HttpInter
 use crate::secrets::SecretsStore;
 use crate::tools::tool::{Tool, ToolDiscoverySummary, ToolError, ToolOutput};
 use crate::tools::wasm::capabilities::Capabilities;
-use crate::tools::wasm::credential_injector::{
+use crate::tools::credentials::{
     InjectedCredentials, host_matches_pattern, inject_credential,
 };
 use crate::tools::wasm::error::WasmError;
